requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
pycodestyle==2.9.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
